+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Release 1.14.0
+## General
+- Added Global Mute toggle to Meadow Remix's "Gameplay" tab
 ## Arena
 - Added import/export map playlists from clipboard
 - Added configurable Artificer explosion capacity